Resolves #40
