+++ src/JS.Abp.DynamicMenu.Blazor/Menus/DynamicMenuMenuContributor.cs
@@ -14,17 +14,16 @@ namespace JS.Abp.DynamicMenu.Blazor.Menus;
 
 public class DynamicMenuMenuContributor : IMenuContributor
 {
-    private readonly IConfiguration _configuration;
-    public DynamicMenuMenuContributor(IConfiguration configuration)
+    public DynamicMenuMenuContributor()
     {
-        _configuration = configuration;
+        
     }
+    
     public async Task ConfigureMenuAsync(MenuConfigurationContext context)
     {
         if (context.Menu.Name == StandardMenus.Main)
         {
             await ConfigureMainMenuAsync(context);
-            await AddDynamicMenuManagementMenuItemAsync(context);
         }
     }
 
@@ -45,90 +44,5 @@ public class DynamicMenuMenuContributor : IMenuContributor
 
         return Task.CompletedTask;
     }
-
-    protected virtual async Task AddDynamicMenuManagementMenuItemAsync(MenuConfigurationContext context)
-    {
-        var l = context.GetLocalizer<DynamicMenuResource>();
-        var dynamicMenuEnabled = _configuration["App:DisableDynamicMenu"];
-        if (dynamicMenuEnabled.IsNullOrWhiteSpace() || !bool.Parse(dynamicMenuEnabled))
-        {
-            var menuAppService = context.ServiceProvider.GetRequiredService<IMenuItemsAppService>();
-
-            var menuItems = await menuAppService.GetListAsync();
-        
-            if (menuItems.Items.Count > 0)
-            {
-                foreach (var menuItemDto in menuItems.Items.Where(x => x.ParentId == null && x.IsActive).OrderBy(x => x.Order)
-                             .ThenBy(x => x.Name))
-                {
-               
-                    var menuItem = context.Menu.FindMenuItem(menuItemDto.Name);
-                    if (menuItem==null)
-                    {
-                        AddChildItems(menuItemDto, menuItems.Items.ToList(), l, context.Menu);
-                    }
-                    else
-                    {
-                        if (!menuItemDto.Url.IsNullOrWhiteSpace())
-                        {
-                            menuItem.Url = menuItemDto.Url;
-                        }
-                        if (!menuItemDto.Icon.IsNullOrWhiteSpace())
-                        {
-                            menuItem.Icon = menuItemDto.Icon;
-                        }
-                        menuItem.Order = menuItemDto.Order;
-                    }
-                }
-            }
-        }
-       
-    }
-
-    private void AddChildItems(MenuItemDto menuItem, List<MenuItemDto> source, IStringLocalizer localizer, IHasMenuItems parent = null)
-    {
-        menuItem.DisplayName = localizer[menuItem.DisplayName];
-        var applicationMenuItem = CreateApplicationMenuItem(menuItem);
-
-        foreach (var item in source.Where(x => x.ParentId == menuItem.Id && x.IsActive).OrderBy(x => x.Order)
-                     .ThenBy(x => x.Name))
-        {
-            AddChildItems(item, source, localizer, applicationMenuItem);
-        }
-
-        parent?.Items.Add(applicationMenuItem);
-    }
-    private ApplicationMenuItem CreateApplicationMenuItem(MenuItemDto menuItem)
-    {
-        if (string.IsNullOrWhiteSpace(menuItem.Permission))
-        {
-            return new ApplicationMenuItem(
-            menuItem.Name,
-            menuItem.DisplayName,
-            menuItem.Url,
-            menuItem.Icon,
-            menuItem.Order,
-            menuItem.Target,
-            menuItem.ElementId,
-            menuItem.CssClass
-        );
-        }
-        else
-        {
-            return new ApplicationMenuItem(
-           menuItem.Name,
-           menuItem.DisplayName,
-           menuItem.Url,
-           menuItem.Icon,
-           menuItem.Order,
-           menuItem.Target,
-           menuItem.ElementId,
-           menuItem.CssClass,
-           requiredPermissionName: menuItem.Permission
-       );
-        }
-
-    }
-
 
 }
